+++ src/screens/Article/Profile.tsx
@@ -5,19 +5,15 @@ import styles from './Profile.style';
 import { typo } from '@/styles';
 
 // TODO: change input type
-function Profile_Chat({ dummyArticle }: any): JSX.Element {
+function Profile({ dummyArticle }: any): JSX.Element {
   return (
-    <View style={styles.userContainer}>
-      <View style={styles.profile}>
-        <Image
-          style={styles.profileImg}
-          source={{ uri: dummyArticle.writer.picture }}
-        />
-        <Text style={{ ...typo.semiTitle }}>
-          {dummyArticle.writer.nickname}
-        </Text>
-      </View>
+    <View style={styles.profile}>
+      <Image
+        style={styles.profileImg}
+        source={{ uri: dummyArticle.writer.picture }}
+      />
+      <Text style={{ ...typo.semiTitle }}>{dummyArticle.writer.nickname}</Text>
     </View>
   );
 }
-export default Profile_Chat;
+export default Profile;
