+++ src/screens/Article/Title.tsx
@@ -6,7 +6,7 @@ import { palette, typo } from '@/styles';
 import { calcTimeDiff } from '@/helpers/functions/index';
 
 // will change input type
-function Title_Info({ dummyArticle }: any): JSX.Element {
+function Title({ dummyArticle }: any): JSX.Element {
   // 남은 시간
   const timeLeft = useMemo(
     () => calcTimeDiff(dummyArticle.created_at, dummyArticle.dueDate),
@@ -19,22 +19,20 @@ function Title_Info({ dummyArticle }: any): JSX.Element {
   );
 
   return (
-    <View style={styles.bigContainer}>
-      <View style={styles.subContainer}>
-        <View style={styles.subConNoBorder}>
-          <Label
-            style={[styles.label, { color: palette.yellow, ...typo.bigTitle }]}
-          >
-            판매
-          </Label>
-          <Text style={{ ...typo.bigTitle }}>{dummyArticle.title}</Text>
-        </View>
-        <View style={[styles.subConNoBorder, { paddingLeft: 15 }]}>
-          <Text style={styles.subText}>{timeBefore}분 전 · </Text>
-          <Text style={styles.subText}>{timeLeft}일 남음</Text>
-        </View>
+    <View style={styles.subContainer}>
+      <View style={styles.subConNoBorder}>
+        <Label
+          style={[styles.label, { color: palette.yellow, ...typo.bigTitle }]}
+        >
+          판매
+        </Label>
+        <Text style={{ ...typo.bigTitle }}>{dummyArticle.title}</Text>
+      </View>
+      <View style={[styles.subConNoBorder, { paddingLeft: 15 }]}>
+        <Text style={styles.subText}>{timeBefore}분 전 · </Text>
+        <Text style={styles.subText}>{timeLeft}일 남음</Text>
       </View>
     </View>
   );
 }
-export default Title_Info;
+export default Title;
