+++ src/screens/Article/TitleInfo.style.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+import { palette } from '@/styles';
+
+const TitleInfoStyles = StyleSheet.create({
+  bigContainer: {
+    borderBottomWidth: 1,
+    borderBottomColor: palette.borderGray,
+  },
+});
+
+export default TitleInfoStyles;
